Revert "SFDM-3384"
This reverts commit 62f23688bac9ee35ecd21e2df07e71b584b6fddc.

+++ src/main/java/ru/beeline/cxbackend/domain/bi/BIInCJStep.java
@@ -7,7 +7,6 @@ package ru.beeline.cxbackend.domain.bi;
 import lombok.*;
 
 import javax.persistence.*;
-import java.math.BigDecimal;
 
 @Builder
 @Getter
@@ -32,7 +31,7 @@ public class BIInCJStep {
     private Long cjStepId;
 
     @Column(name = "\"order\"")
-    private BigDecimal order;
+    private Long order;
 
     @Column(name = "id_bi", insertable = false, updatable = false)
     private Long biId;
