+++ src/main/java/ru/beeline/cxbackend/domain/cj/CJStep.java
@@ -8,7 +8,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.*;
 
 import javax.persistence.*;
-import java.math.BigDecimal;
 
 @Builder
 @Getter
@@ -25,7 +24,7 @@ public class CJStep {
     private Long id;
 
     @Column(name = "\"order\"")
-    private BigDecimal order;
+    private Integer order;
 
     private String name;
