+++ src/main/java/ru/beeline/cxbackend/dto/BiByCjStepDto.java
@@ -7,11 +7,9 @@ package ru.beeline.cxbackend.dto;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class BiByCjStepDto {
     @JsonProperty("id_bi")
     private Long idBi;
-    private BigDecimal order;
+    private Long order;
 }
