+++ src/main/java/ru/beeline/cxbackend/dto/CjStepDto.java
@@ -6,11 +6,9 @@ package ru.beeline.cxbackend.dto;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class CjStepDto {
     private String name;
-    private BigDecimal order;
+    private Integer order;
     private String description;
 }
