+++ src/main/java/ru/beeline/cxbackend/dto/CjStepFullDto.java
@@ -10,8 +10,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -20,7 +18,7 @@ public class CjStepFullDto {
 
     private Long id;
     private String name;
-    private BigDecimal order;
+    private Integer order;
     private String description;
     @JsonProperty("id_cj")
     private Long cjId;
