+++ src/main/java/ru/beeline/cxbackend/dto/StepDto.java
@@ -8,7 +8,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 import lombok.Getter;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -23,7 +22,7 @@ public class StepDto {
     @JsonProperty("id_cj")
     private Long cjId;
 
-    private BigDecimal order;
+    private Integer order;
 
     private String name;
