+++ src/main/java/ru/beeline/cxbackend/dto/StepDtoV2.java
@@ -9,7 +9,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -25,7 +24,7 @@ public class StepDtoV2 {
 
     private Long cjId;
 
-    private BigDecimal order;
+    private Integer order;
 
     private String name;
