+++ src/main/java/ru/beeline/cxbackend/dto/StepDtoV3.java
@@ -5,7 +5,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -21,7 +20,7 @@ public class StepDtoV3 {
 
     private Long cjId;
 
-    private BigDecimal order;
+    private Integer order;
 
     private String name;
